backend/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
psycopg==3.1.18
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
